% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Data_AddTimeTrace.R
\name{AddTimeTrace}
\alias{AddTimeTrace}
\title{Add or replace the TimeTrace of an EPhysContainer}
\usage{
AddTimeTrace(x, from, to, by, Unit)
}
\arguments{
\item{x}{An \code{EPhysContainer}.}

\item{from, to}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{by}{number: increment of the sequence.}

\item{Unit}{Character. A valid udunits time unit symbol (e.g., \code{"s"}, \code{"ms"}).}
}
\value{
The modified \code{EPhysContainer} object.
}
\description{
Fills the \code{TimeTrace} slot using \code{base::seq()} and sets the
\code{TimeUnits} slot after validating the unit symbol.
}
\details{
\code{Unit} is checked by attempting \code{units::set_units(1, Unit, mode = "standard")}.
The resulting \code{TimeTrace} must be strictly increasing (i.e., \code{diff(TimeTrace) > 0}).
}
\examples{
# Add a 0..1 s trace in 1 ms steps:
# obj <- AddTimeTrace(obj, from = 0, to = 1, by = 0.001, Unit = "s")
}
\seealso{
\code{\link[base]{seq}}
}
