% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_General_Bin.R
\name{Bin}
\alias{Bin}
\title{Bin spikes into an EPhysContinuous object}
\usage{
Bin(X, binWidth, StimulusBinFunction = mean, parallel = !interactive())
}
\arguments{
\item{X}{An \code{EPhysEvents} object.}

\item{binWidth}{A \pkg{units} object specifying the width of each time bin;
its unit will be set to \code{TimeUnits(X)}.}

\item{StimulusBinFunction}{Function to bin stimulus trace by (if present).}

\item{parallel}{logical.
For \code{EPhysEvents} / \code{EPhysContinuous} methods only:
if \code{TRUE}, use \code{future.apply::future_lapply()} over runs.}
}
\value{
An \code{EPhysContinuous} object with:
\describe{
  \item{Data}{A 3D numeric array \code{[time × trial × channel]} of spike counts.}
  \item{TimeTrace}{Numeric vector of bin starts (in \code{TimeUnits(X)}).}
  \item{Channels}{Character vector of channel names.}
  \item{StimulusTrace}{Empty numeric vector.}
  \item{TimeUnits}{Copied from \code{TimeUnits(X)}.}
  \item{Metadata}{Copied from \code{Metadata(X)}.}
}
}
\description{
Bins per‐trial spike times from an \code{EPhysEvents} object into a
left‐inclusive, right‐exclusive 3D numeric count array
(\code{time × trial × channel}), storing counts and metadata.
}
