% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_General_ComputeIEI.R
\name{ComputeIEI}
\alias{ComputeIEI}
\alias{ComputeIEI,EPhysEvents-method}
\title{Compute inter‐event‐intervals (IEI) from an EPhysEvents object}
\usage{
ComputeIEI(x)

\S4method{ComputeIEI}{EPhysEvents}(x)
}
\value{
An \code{EPhysIEI} object with:
\describe{
  \item{Metadata}{Copied from \code{x@Metadata}.}
  \item{Data}{A list of length \code{nrow(x@Metadata)}, each element a named list of
    channels. For each channel, a data.frame with columns:
    \itemize{
      \item \code{Time}: original zero‐based timestamps,
      \item \code{IEI}: inter‐event intervals (\code{c(NA, diff(Time))}).
    }
  }
}
}
\description{
Transforms each channel’s timestamp vectors in an \code{EPhysEvents} object
into a two‐column data.frame of \code{Time} and \code{IEI}, and returns an
\code{EPhysIEI} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ComputeIEI(EPhysEvents)}: Method for EPhysEvents

}}
\examples{
\dontrun{
rec_ev <- LoadEPhysEvents("data/session1", "types.tsv")
rec_iei <- ComputeIEI(rec_ev)
str(rec_iei@Data[[1]]$ChannelA)
}
}
