% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_General_StepStimMetrics.R
\name{count_in}
\alias{count_in}
\title{left Inclusive, right exclusive window counting}
\usage{
count_in(x, a, b)
}
\arguments{
\item{a, b}{Left () and right margins of window}
}
\description{
left Inclusive, right exclusive window counting
}
\keyword{internal}
