% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Data_Merge.R
\name{Merge}
\alias{Merge}
\title{Merge two EPhysContainer objects along channels}
\usage{
Merge(x, y, match_by = c("Step", "Experiment"))
}
\arguments{
\item{x, y}{\code{EPhysContainer} objects.}

\item{match_by}{Character vector of column names in the Runs (Metadata) used
to align Runs. Defaults to \code{c("Step","Experiment")}.}
}
\value{
A merged \code{EPhysContainer}.
}
\description{
Concatenate channels from two \code{EPhysContainer} objects after aligning
**Runs** (rows in \code{Metadata}) by key columns given in \code{match_by}
(default \code{c("Step","Experiment")}). Supports two data layouts:
(1) list-of-arrays where the 3rd dim is channels, and
(2) list-of-lists where the 2nd-level list enumerates channels.
Both inputs must use the same layout.
}
